% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSignatures.R
\name{alignmentSignature}
\alias{alignmentSignature}
\title{From the list of Parameters from the getPMSignature function, 
return the order of signatures so that signatures from the several lists are well aligned.}
\usage{
alignmentSignature(Params)
}
\arguments{
\item{Params}{the vector of the result of getPMSignature (EstimatedParameters class)}
}
\description{
From the list of Parameters from the getPMSignature function, 
return the order of signatures so that signatures from the several lists are well aligned.
}

