% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boundaryTurbo_F}
\alias{boundaryTurbo_F}
\title{Check whether the parameter F is within the appropriate range}
\usage{
boundaryTurbo_F(turboF, fdim, signatureNum)
}
\arguments{
\item{turboF}{F (converted for turboEM)}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}

\item{signatureNum}{the number of mutation signatures}
}
\description{
Check whether the parameter F is within the appropriate range
}

