% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boundaryTurbo_Q}
\alias{boundaryTurbo_Q}
\title{Check whether the parameter Q is within the appropriate range}
\usage{
boundaryTurbo_Q(turboQ, signatureNum, sampleNum)
}
\arguments{
\item{turboQ}{Q (converted for turboEM)}

\item{signatureNum}{the number of mutation signatures}

\item{sampleNum}{the number of cancer genomes}
}
\description{
Check whether the parameter Q is within the appropriate range
}

