% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSignature.R
\name{calcPMSLikelihood}
\alias{calcPMSLikelihood}
\title{A function for calculating the log-likelihood from the data and parameters}
\usage{
calcPMSLikelihood(p, y)
}
\arguments{
\item{p}{this variable includes the parameters for mutation signatures and membership parameters}

\item{y}{this variable includes the information on the mutation features,
the number of mutation signatures specified and so on}
}
\description{
A function for calculating the log-likelihood from the data and parameters
}

