% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{convertFromTurbo_F}
\alias{convertFromTurbo_F}
\title{Restore the converted parameter F for turboEM}
\usage{
convertFromTurbo_F(turboF, fdim, signatureNum, isBackground)
}
\arguments{
\item{turboF}{F (converted for turboEM)}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}

\item{signatureNum}{the number of mutation signatures}

\item{isBackground}{the logical value showing whether a background mutaiton features is included or not}
}
\description{
Restore the converted parameter F for turboEM
}

