% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{convertSignatureMatrixToVector}
\alias{convertSignatureMatrixToVector}
\title{the function for converting the mutation signature matrix to a vector}
\usage{
convertSignatureMatrixToVector(Fmat, fdim)
}
\arguments{
\item{Fmat}{a matrix for mutation signature}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}
}
\description{
the function for converting the mutation signature matrix to a vector
}

