% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertToTurbo_F}
\alias{convertToTurbo_F}
\title{Convert the parameter F so that turboEM can treat}
\usage{
convertToTurbo_F(vF, fdim, signatureNum, isBackground)
}
\arguments{
\item{vF}{F (converted to a vector)}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}

\item{signatureNum}{the number of mutation signatures}

\item{isBackground}{the logical value showing whether a background mutaiton features is included or not}
}
\description{
Convert the parameter F so that turboEM can treat
}
