% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertToTurbo_Q}
\alias{convertToTurbo_Q}
\title{Convert the parameter Q so that turboEM can treat}
\usage{
convertToTurbo_Q(vQ, signatureNum, sampleNum)
}
\arguments{
\item{vQ}{Q (converted to a vector)}

\item{signatureNum}{the number of mutation signatures}

\item{sampleNum}{the number of cancer genomes}
}
\description{
Convert the parameter Q so that turboEM can treat
}

