% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{EstimatedParameters-class}
\alias{EstimatedParameters-class}
\title{An S4 class representing the estimated parameters}
\description{
An S4 class representing the estimated parameters
}
\section{Slots}{

\describe{
\item{\code{sampleList}}{a list of sample names observed in the input mutation data}

\item{\code{signatureNum}}{the number of mutation signatures specified at the time of estimation}

\item{\code{isBackGround}}{the flag showing whether the background signature data is used or not.}

\item{\code{signatureFeatureDistribution}}{estimated parameters for mutation signatures}

\item{\code{sampleSignatureDistribution}}{estimated parameters for memberships of mutation signatures for each sample}

\item{\code{loglikelihood}}{the loglikelihood about the estimated parameters}
}}

