% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{getCosDistance}
\alias{getCosDistance}
\title{the function for calculating cosine distances between two mutation signature matrices.}
\usage{
getCosDistance(F_1, F_2, fdim)
}
\arguments{
\item{F_1}{the first matrix for mutation signature}

\item{F_2}{the second matrix for mutation signature}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}
}
\description{
the function for calculating cosine distances between two mutation signature matrices.
}

