% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{getLogLikelihoodC}
\alias{getLogLikelihoodC}
\title{Calculate the value of the log-likelihood for given parameters}
\usage{
getLogLikelihoodC(vG, vF, vQ, fdim, signatureNum, sampleNum, patternNum,
  isBackground, vF0)
}
\arguments{
\item{vG}{the matrix of mutation feature data}

\item{vF}{the parameter F (converted to a vector)}

\item{vQ}{the parameter Q (converted to a vector)}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}

\item{signatureNum}{the number of mutation signatures}

\item{sampleNum}{the number of cancer genomes}

\item{patternNum}{the number of possible combinations of all the mutation features}

\item{isBackground}{the logical value showing whether a background mutaiton features is included or not}

\item{vF0}{a background mutaiton features}
}
\description{
Calculate the value of the log-likelihood for given parameters
}

