% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class.R
\docType{methods}
\name{getMembershipValue,EstimatedParameters-method}
\alias{getMembershipValue,EstimatedParameters-method}
\title{Get the values of estimated membership parameters}
\usage{
\S4method{getMembershipValue}{EstimatedParameters}(object)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of getPMSignature).}
}
\description{
Get the values of estimated membership parameters
}

