% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{getMembershipValue}
\alias{getMembershipValue}
\title{Get the values of estimated membership parameters}
\usage{
getMembershipValue(object)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of getPMSignature).}
}
\value{
membership parameter values estimated by \code{getPMSignature}
}
\description{
Get the values of estimated membership parameters
}
\examples{
After obtaining EstimatedParameters (typically by \\code{getPMSignature}) as Param,
print(getMembershipValue(Param))

}
