% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFile.R
\name{getMutationFeatureVector}
\alias{getMutationFeatureVector}
\title{get mutation feature vector from context sequence data and reference and alternate allele information}
\usage{
getMutationFeatureVector(context, ref_base, alt_base, strandInfo = NULL,
  numBases, type)
}
\arguments{
\item{context}{the context sequence data around the mutated position. This shoud be Biostrings::DNAStringSet class}

\item{ref_base}{the reference bases at the mutated position.}

\item{alt_base}{the alternate bases at the mutated position.}

\item{strandInfo}{transcribed strand information at the mutated position. (this is optional)}

\item{numBases}{the number of flanking bases around the mutated position.}

\item{type}{the type of mutation feature vecotr (should be "independent" or "full").}
}
\description{
get mutation feature vector from context sequence data and reference and alternate allele information
}

