% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignature.R
\name{getMutMembership}
\alias{getMutMembership}
\title{Obtain the standard error estimates for parameters for mutation signatures and memberships}
\usage{
getMutMembership(MutationFeatureData, EstimatedParameters)
}
\arguments{
\item{EstimatedParameters}{the estimated parameters (EstimatedParameter class (S4 class)) by the \code{getPMSignature} function.}

\item{mutationFeatureData}{the mutation data (MutationFeatureData class (S4 class)) by the \code{readMPFile} or \code{readMFVFile} functions.}
}
\value{
a data frame of mutation wise membership.
}
\description{
Obtain the standard error estimates for parameters for mutation signatures and memberships
}
\examples{
After obtaining mutationFeatureData (see e.g., by \\code{readMPFile} function) as G, 
and EstimatedParameters (e.g., by \\code{getPMSignature} function) as Param,
mutMembership <- getMutMembership(G, Param)

}

