% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPMSignature}
\alias{getPMSignature}
\title{Obtain the parameters for mutation signatures and memberships}
\usage{
getPMSignature(G, K, fdim, isBG, BG0, numInit)
}
\arguments{
\item{G}{the matrix of mutation feature data}

\item{K}{the number of mutation signatures}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}

\item{isBG}{the logical value showing whether a background mutaiton features is included or not}

\item{BG0}{a background mutaiton features}

\item{numInit}{the number of performing calculations with different initial values}
}
\description{
Obtain the parameters for mutation signatures and memberships
}

