% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareSignatures.R
\name{getSignaturesForMultipleK}
\alias{getSignaturesForMultipleK}
\title{Generate the figure of mutation signature for multiple K aligning the signatures.}
\usage{
getSignaturesForMultipleK(inputFile, numBases = 5, trDir = TRUE,
  startK = NULL, endK = 6)
}
\arguments{
\item{inputFile}{the path for the input file (MP format)}

\item{numBases}{the number of flanking bases}

\item{trDir}{the index for the usage of transcription strand bias}

\item{startK}{the minimum number of signatures}

\item{endK}{the maximum number of signatures}
}
\description{
Generate the figure of mutation signature for multiple K aligning the signatures.
}

