% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.R
\docType{methods}
\name{getSignatureValue,EstimatedParameters,numeric-method}
\alias{getSignatureValue,EstimatedParameters,numeric-method}
\title{Get the values of estimated signatures}
\usage{
\S4method{getSignatureValue}{EstimatedParameters,numeric}(object, sinInd = 1)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of getPMSignature).}

\item{sigInd}{the index of the estimated signatures}
}
\description{
Get the values of estimated signatures
}

