% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class.R
\name{getSignatureValue}
\alias{getSignatureValue}
\title{Get the values of estimated signatures}
\usage{
getSignatureValue(object, sinInd = 1)
}
\arguments{
\item{object}{the EstimatedParameters class (the result of getPMSignature)}

\item{sigInd}{the index of the estimated signatures}
}
\value{
mutation signature parameter values estimated by \code{getPMSignature}
}
\description{
Get the values of estimated signatures
}
\examples{
After obtaining EstimatedParameters (typically by \\code{getPMSignature}) as Param,
print(getSignatureValue(Param, 1))
}

