% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{makeSimData}
\alias{makeSimData}
\title{Generate simulation data}
\usage{
makeSimData(type = "independent", numBases = 3, trDir = FALSE,
  fdim = NULL, K = 3, sampleNum = 10, mutationNum = 100,
  param_alpha = 1, param_gamma = 1, isBG = FALSE)
}
\arguments{
\item{type}{this argument can take either independent, full}

\item{numBases}{the number of upstream and downstream flanking bases}

\item{trDir}{the index representing whether transcription direction is considered or not}

\item{K}{the number of mutation signature}

\item{sampleNum}{the number of cancer genomes in the simulation data}

\item{mutationNum}{the number of mutations in each cancer genome}

\item{param_alpha}{the parameter of the Dirichlet distribution for the sample signature distribution}

\item{param_gamma}{the parameter of the Diriculet distribution for signature feature distribution}

\item{isBG}{the logical value showing whether the background signature is used or not}
}
\description{
Generate simulation data
}

