% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{MutationFeatureData-class}
\alias{MutationFeatureData-class}
\title{An S4 class representing the mutation data}
\description{
An S4 class representing the mutation data
}
\section{Slots}{

\describe{
\item{\code{featureVectorList}}{a list of feature vectors actually observed in the input mutation data}

\item{\code{sampleList}}{a list of sample names observed in the input mutation data}

\item{\code{countData}}{a matrix representing the number of mutations and samples.
The (first, second, third) colums are for (mutation pattern indice, sample indice, frequencies).}
}}

