% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSignature.R
\name{mySquareEM}
\alias{mySquareEM}
\title{A function for estimating parameters using Squared EM algorithm}
\usage{
mySquareEM(p, y, tol = 1e-04, maxIter = 10000)
}
\arguments{
\item{p}{this variable includes the parameters for mutation signatures and membership parameters}

\item{y}{this variable includes the information on the mutation features,
the number of mutation signatures specified and so on}

\item{tol}{tolerance for the estimation
(when the difference of log-likelihoods become below this value, stop the estimation)}

\item{maxIter}{the maximum number of iteration of estimation}
}
\description{
A function for estimating parameters using Squared EM algorithm
}

