% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PMSboundary}
\alias{PMSboundary}
\title{A functional for generating the function checking the parameter (p) is within the restricted conditions or not}
\usage{
PMSboundary(y)
}
\arguments{
\item{y}{this variable includes the information on the mutation features,
the number of mutation signatures specified and so on}
}
\description{
A functional for generating the function checking the parameter (p) is within the restricted conditions or not
}

