% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readFile.R
\name{readBGFile}
\alias{readBGFile}
\title{Read and format the background vector data}
\usage{
readBGFile(mutationFeatureData)
}
\arguments{
\item{mutationFeatureData}{the mutation data processed in the function (readMutFile or readRawMutfeatFile)}
}
\description{
Read and format the background vector data
}

