% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readFile.R
\name{readMPFile}
\alias{readMPFile}
\title{Read the raw mutation data of mutation position format.}
\usage{
readMPFile(infile, numBases = 3, trDir = FALSE, type = "independent")
}
\arguments{
\item{infile}{the path for the input file for the mutation data.}

\item{numBases}{the number of upstream and downstream flanking bases
(including the mutated base) to take into account.}

\item{trDir}{the index representing whether transcription direction is considered or not}

\item{type}{this argument can take either independent, full, or custom.}
}
\description{
Read the raw mutation data of mutation position format.
}

