% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationRatio.R
\name{readMPFile_localMutRate}
\alias{readMPFile_localMutRate}
\title{read the mutation position format file and extract the genomic position as a mutation feature}
\usage{
readMPFile_localMutRate(infile)
}
\arguments{
\item{infile}{the path for the input mutation position format file}
}
\description{
read the mutation position format file and extract the genomic position as a mutation feature
}

