% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readMutFile}
\alias{readMutFile}
\title{Read the raw mutation data.}
\usage{
readMutFile(infile, numBases = 3, trDir = FALSE)
}
\arguments{
\item{infile}{the path for the input file for the mutation data.}

\item{numBases}{the number of upstream and downstream flanking bases
(including the mutated base) to take into account.}
}
\description{
Read the raw mutation data.
}

