% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readRawMutfeatFile}
\alias{readRawMutfeatFile}
\title{Read the raw mutation data with the mutation feature format.}
\usage{
readRawMutfeatFile(infile)
}
\arguments{
\item{infile}{the path for the input file for the mutation feature data.}
}
\description{
Read the raw mutation data with the mutation feature format.
}

