% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{readRawMutFile}
\alias{readRawMutFile}
\title{Read and format the raw mutation data}
\usage{
readRawMutFile(infile)
}
\arguments{
\item{infile}{the path for the input file for the raw mutation data}
}
\description{
Read and format the raw mutation data
}

