% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateMstepFC}
\alias{updateMstepFC}
\title{Update the parameter F (M-step in the EM-algorithm)}
\usage{
updateMstepFC(vTheta, vG, fdim, signatureNum, sampleNum, patternNum,
  isBackground)
}
\arguments{
\item{vTheta}{theparameter Theta (converted to a vector)}

\item{vG}{the matrix of mutation feature data}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}

\item{signatureNum}{the number of mutation signatures}

\item{sampleNum}{the number of cancer genomes}

\item{patternNum}{the number of possible combinations of all the mutation features}

\item{isBackground}{the logical value showing whether a background mutaiton features is included or not}
}
\description{
Update the parameter F (M-step in the EM-algorithm)
}

