% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateMstepFQC}
\alias{updateMstepFQC}
\title{Update the parameter F and Q (M-step in the EM-algorithm)}
\usage{
updateMstepFQC(vPatternList, vSparseCount, nTheta, fdim, signatureNum,
  sampleNum, patternNum, samplePatternNum, isBackground)
}
\arguments{
\item{vPatternList}{The list of possible mutation features (converted to a vector)}

\item{vSparseCount}{The table showing (mutation feature, sample, the number of mutation) (converted to a vector)}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}

\item{signatureNum}{the number of mutation signatures}

\item{sampleNum}{the number of cancer genomes}

\item{patternNum}{the number of possible combinations of all the mutation features}

\item{samplePatternNum}{the number of possible combination of samples and mutation patternns}

\item{isBackground}{the logical value showing whether a background mutaiton features is included or not}

\item{vF}{F (converted to a vector)}

\item{vQ}{Q (converted to a vector)}

\item{vF0}{a background mutaiton features}
}
\description{
Update the parameter F and Q (M-step in the EM-algorithm)
}

