% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateMstepQC}
\alias{updateMstepQC}
\title{Update the parameter Q (M-step in the EM-algorithm)}
\usage{
updateMstepQC(vTheta, vG, signatureNum, sampleNum, patternNum)
}
\arguments{
\item{vTheta}{theparameter Theta (converted to a vector)}

\item{vG}{the matrix of mutation feature data}

\item{signatureNum}{the number of mutation signatures}

\item{sampleNum}{the number of cancer genomes}

\item{patternNum}{the number of possible combinations of all the mutation features}
}
\description{
Update the parameter Q (M-step in the EM-algorithm)
}

