% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updatePMSParam}
\alias{updatePMSParam}
\title{A function for updating parameters using EM-algorithm}
\usage{
updatePMSParam(p, y)
}
\arguments{
\item{p}{this variable includes the parameters for mutation signatures and membership parameters}

\item{y}{this variable includes the information on the mutation features,
the number of mutation signatures specified and so on}
}
\description{
A function for updating parameters using EM-algorithm
}

