% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualizeSignature.R
\docType{methods}
\name{visCorrelation}
\alias{visCorrelation}
\title{Visualize Correlation}
\usage{
visCorrelation(object)
}
\arguments{
\item{object}{EstimatedParameters object.}
}
\value{
ggplot object.
}
\description{
Visualize pairwise signature correlations using a heatmap.
}

