% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualizeSignature.R
\name{visMembership}
\alias{visMembership}
\title{Visualize estimated membership parameters}
\usage{
visMembership(object1, object2, ylog = FALSE, sortSampleNum = TRUE,
  multiplySampleNum = TRUE, fromSample = NULL, toSample = NULL,
  reorderSig = NULL, colourBrewer = NULL)
}
\arguments{
\item{object1}{MutationFeatureData class}

\item{object2}{EstimatedParameters class}

\item{ylog}{barplot height is an logarithms of #mutaitons or not
(defalt: FALSE, active only if multiplySampleNum = TRUE).}

\item{sortSampleNum}{samples are sorted according to #mutations (default: TRUE).}

\item{multiplySampleNum}{barplot height for each sample is multiplied by #mutations (default: TRUE).}

\item{fromSample}{only samples from the specified index will be plotted (default: NULL).}

\item{reorderSig}{the order of signatures are reordered according to the specified order (default: NULL).}

\item{colourBrewer}{NULL}

\item{fromSample}{only samples until the specified index will be plotted (default: NULL).}
}
\value{
a figure of estimaged membership parameter via ggplot2 is generated
(therefore, can be saved using \code{ggsave} function).
}
\description{
Visualize estimated membership parameters
}
\examples{
After obtaining EstimatedParameters (typically by \\code{getPMSignature}) as Param,
visPMSignature(G, Param)

You can equate the heights of barplot.
visPMSignature(G,Param, multiplySampleNum = TRUE)
}

