% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeSignature2.R
\name{visMembership2}
\alias{visMembership2}
\title{Visualize estimated membership parameters using ggvis}
\usage{
visMembership2(object)
}
\arguments{
\item{object}{EstimatedParameters class}
}
\value{
a figure of estimaged membership parameter via ggvis is generated
}
\description{
Visualize estimated membership parameters using ggvis
}
\examples{
After obtaining EstimatedParameters (typically by \\code{getPMSignature}) as Param,
visPMSignature2(Param)

}

