% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualizeSignature.R
\name{visPMS_full}
\alias{visPMS_full}
\title{visualize probabisitic mutaiton signature for the full model}
\usage{
visPMS_full(Fvec, numBases, trDir)
}
\arguments{
\item{numBases}{the number of flanking bases}

\item{trDir}{the index showing whether the transcription direction is used or not}

\item{vF}{a vector for mutation signature}
}
\description{
Generate visualization of mutation signatures for the model with
  substitution patterns and flanking bases represented by the full
  representation.
}

