% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visPMS_full5}
\alias{visPMS_full5}
\title{visualize probabisitic mutaiton signature for the full model
with two 5' and 3' bases}
\usage{
visPMS_full5(v1 = rep(1/1536, 1536))
}
\arguments{
\item{v1}{a vector for mutation signature}
}
\description{
Generate visualization of mutation signatures for the model with
  substitution patterns and two 5' and 3' bases represented by the full
  representation.
}

