% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualizeSignature.R
\name{visPMS_ind}
\alias{visPMS_ind}
\title{visualize probabisitic mutaiton signature for the independent model}
\usage{
visPMS_ind(vF, numBases, baseCol = NA, trDir, charSize = 1.2)
}
\arguments{
\item{vF}{a matrix for mutation signature}

\item{numBases}{the number of flanking bases}

\item{centerBases}{the number of flanking bases}
}
\description{
Generate visualization of mutation signatures for the model with
  substitution patterns and flanking bases represented by the indepenent
  representation.
}

