% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visPMS_ind5}
\alias{visPMS_ind5}
\title{visualize probabisitic mutaiton signature}
\usage{
visPMS_ind5(v1 = rep(1/6, 6), V2 = matrix(1/4, 4, 4))
}
\arguments{
\item{v1}{A vector of a substitution pattern.}

\item{v2}{Vectors of two 5' and 3' bases.}
}
\description{
Generate visualization of mutation signatures for the model with
  substitution patterns and two 5' and 3' bases represented by the indepenent
  representation.
}

