% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualizeSignature.R
\name{visPMSignature}
\alias{visPMSignature}
\title{visualize probabisitic mutaiton signature}
\usage{
visPMSignature(object, sinInd = 1, ...)
}
\arguments{
\item{object}{}

\item{sigInd}{the index of the mutation signature plotted}

\item{trDir}{the index whether the strand direction is plotted or not}

\item{baseCol}{the colour of the bases (A, C, G, T, plus strand, minus strand) (active only if type = "independent")}

\item{charSize}{the size of the character (active only if type = "independent")}

\item{isScale}{the index whether the height of the flanking base is changed or not (active only if type = "independent")}

\item{alpha}{the parameter for the Renyi entropy (applicable only if the isScale is TRUE)}
}
\value{
a figure of estimaged probabilistic mutation signature via ggplot2 is generated
(therefore, can be saved using \code{ggsave} function.
}
\description{
test
}
\examples{
test
}

