% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{bootPMSignature}
\alias{bootPMSignature}
\title{Obtain the standard error estimates for parameters for mutation signatures and memberships}
\usage{
bootPMSignature(G, K, fdim, isBG, BG0, F0, Q0, bootNum)
}
\arguments{
\item{G}{the matrix of mutation feature data}

\item{K}{the number of mutation signatures}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}

\item{isBG}{the logical value showing whether a background mutaiton features is included or not}

\item{BG0}{a background mutaiton features}

\item{F0}{the initial value for the parameter of mutation signatures used for bootstraped parameter estimations}

\item{Q0}{the initial value for the parameter of memberships used for bootstraped parameter estimations}

\item{bootNum}{the number of performing bootstrap calculations}
}
\description{
Obtain the standard error estimates for parameters for mutation signatures and memberships
}

