% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertFromTurbo_Q}
\alias{convertFromTurbo_Q}
\title{Restore the converted parameter Q for turboEM}
\usage{
convertFromTurbo_Q(turboQ, signatureNum, sampleNum)
}
\arguments{
\item{turboQ}{Q (converted for turboEM)}

\item{signatureNum}{the number of mutation signatures}

\item{sampleNum}{the number of cancer genomes}
}
\description{
Restore the converted parameter Q for turboEM
}

