% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLogLikelihoodC}
\alias{getLogLikelihoodC}
\title{Calculate the value of the log-likelihood for given parameters}
\usage{
getLogLikelihoodC(vPatternList, vSparseCount, vF, vQ, fdim, signatureNum,
  sampleNum, patternNum, samplePatternNum, isBackground, vF0)
}
\arguments{
\item{vPatternList}{The list of possible mutation features (converted to a vector)}

\item{vSparseCount}{The table showing (mutation feature, sample, the number of mutation) (converted to a vector)}

\item{vF}{F (converted to a vector)}

\item{vQ}{Q (converted to a vector)}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}

\item{signatureNum}{the number of mutation signatures}

\item{sampleNum}{the number of cancer genomes}

\item{patternNum}{the number of possible combinations of all the mutation features}

\item{samplePatternNum}{the number of possible combination of samples and mutation patternns}

\item{isBackground}{the logical value showing whether a background mutaiton features is included or not}

\item{vF0}{a background mutaiton features}
}
\description{
Calculate the value of the log-likelihood for given parameters
}
