% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualizeSignature.R
\name{getMutNum}
\alias{getMutNum}
\title{Obtain somatic mutation count for each sample}
\usage{
getMutNum(object)
}
\arguments{
\item{object}{MutationFeatureData}
}
\value{
the number of somatic mutation stored into the instance of mutationFeatureData class.
}
\description{
Obtain somatic mutation count for each sample
}
\examples{
After obtaining mutationFeatureData (see e.g., readMPFile function) as G,
mutNum <- getMutNum(G)
print(mutNum)
}

