% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSignature.R
\name{getPMSignature}
\alias{getPMSignature}
\title{Obtain the parameters for mutation signatures and memberships}
\usage{
getPMSignature(mutationFeatureData, K, BG = NULL, numInit = 10,
  tol = 1e-04, maxIter = 10000)
}
\arguments{
\item{mutationFeatureData}{the mutation data processed in the function (readMutFile or readRawMutfeatFile)}

\item{K}{the number of mutation signatures}

\item{numInit}{the number of performing calculations with different initial values}

\item{tol}{tolerance for the estimation
(when the difference of log-likelihoods become below this value, stop the estimation)}

\item{maxIter}{the maximum number of iteration of estimation}

\item{isBG}{the logical value showing whether a background mutaiton features is included or not}

\item{BG0}{a background mutaiton features}
}
\description{
Obtain the parameters for mutation signatures and memberships
}

