% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPMSignature}
\alias{getPMSignature}
\title{Obtain the parameters for mutation signatures and memberships}
\usage{
getPMSignature(G, K, isBG = FALSE, BG0 = 0, numInit = 10)
}
\arguments{
\item{G}{the mutation data processed in the function (readMutFile or readRawMutfeatFile)}

\item{K}{the number of mutation signatures}

\item{isBG}{the logical value showing whether a background mutaiton features is included or not}

\item{BG0}{a background mutaiton features}

\item{numInit}{the number of performing calculations with different initial values}
}
\description{
Obtain the parameters for mutation signatures and memberships
}

