% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{MetaInformation-class}
\alias{MetaInformation-class}
\title{An S4 class to represent a mutation meta information common to many data types}
\description{
An S4 class to represent a mutation meta information common to many data types
}
\section{Slots}{

\describe{
\item{\code{type}}{type of data format (independent, full, custom)}

\item{\code{flankingBasesNum}}{the number of flanking bases to consider (only apprecable for independent and full types)}

\item{\code{transcriptionDirection}}{the flag representing whether transcription direction is considered or not}

\item{\code{possibleFeatures}}{a vector representing the numbers of possible values for each mutation feature}
}}

