% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readBGFile}
\alias{readBGFile}
\title{Read and format the background vector data}
\usage{
readBGFile(bgfile)
}
\arguments{
\item{bgfile}{the path for the background mutation signature file}
}
\description{
Read and format the background vector data
}

