% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFile.R
\name{readBGFile}
\alias{readBGFile}
\title{Read and format the background vector data}
\usage{
readBGFile(mutationFeatureData)
}
\arguments{
\item{mutationFeatureData}{the mutation data processed in the function (\code{readMPFile} or \code{readMFVFile})}
}
\value{
The output is a background probability vector corresponding to the input mutation data.
}
\description{
This function provides a background probability for each mutation feature,
that is available only when the model type is "independent" or "full",
and the numBases is either of 3, 5, 7 or 9.

The background probability vectors are calculated by the function \code{getBackgroudSignature},
checking the frequencies of consecutive nucleotides on \strong{exonic regions}.
Therefore, when you are using whole genome sequencing data,
using the background signatures provided by this function may not be appropriate.
}
\examples{
inputFile <- system.file("extdata/Nik_Zainal_2012.mutationPositionFormat.txt.gz", package="pmsignature")
G <- readMPFile(inputFile, numBases = 7)
BG_prob <- readBGFile(G)
}

