% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{readMPFile}
\alias{readMPFile}
\title{Read the raw mutation data of Mutation Position Format.}
\usage{
readMPFile(infile, numBases = 3, trDir = FALSE, type = "independent",
  bs_genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
  txdb_transcript = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene)
}
\arguments{
\item{infile}{the path for the input file for the mutation data of Mutation Position Format.}

\item{numBases}{the number of upstream and downstream flanking bases
(including the mutated base) to take into account.}

\item{trDir}{the index representing whether transcription direction is considered or not.
The gene annotation information is given by UCSC knownGene (TxDb.Hsapiens.UCSC.hg19.knownGene object).
When trDir is TRUE, the mutations located in intergenic region are excluded from the analysis.}

\item{type}{this argument can take either "independent", "full", or "custom".}

\item{bs_genome}{this argument specifies the reference genome (e.g., BSgenome.Mmusculus.UCSC.mm10 can be used for the mouse genome).
See https://bioconductor.org/packages/release/bioc/html/BSgenome.html for the available genome list}

\item{txdb_transcript}{this argument specified the transcript database (e.g., TxDb.Mmusculus.UCSC.mm10.knownGene can be used for the mouse genome).
See https://bioconductor.org/packages/release/bioc/html/AnnotationDbi.html for details.}
}
\value{
The output is an instance of MutationFeatureData S4 class (which stores summarized information on mutation data).
This will be typically the input of \code{getPMSignature} function for estimating the parameters of mutation signatures and memberships.
}
\description{
The mutation position format is tab-delimited text file, where 
the 1st-5th columns shows sample names, chromosome names, 
coordinates, reference bases (A, C, G, or T) and
the alternate bases (A, C, G, or T), respectively. An example is as follows;

---

sample1 chr1 100 A C

sample1 chr1 200 A T

sample1 chr2 100 G T

sample2 chr1 300 T C

sample3 chr3 400 T C

---

Also, this function usually can accept compressed files (e.g., by gzip, bzip2 and so on) when using recent version of R.
Currently, only UCSC hg19 (BSgenome.Hsapiens.UCSC.hg19) is supported.
}
\examples{
We have example data in the pmsignature package;
inputFile <- system.file("extdata/Nik_Zainal_2012.mutationPositionFormat.txt.gz", package="pmsignature")
G <- readMPFile(inputFile, numBases = 5)

When assuming non-independent model;
G <- readMPFile(inputFile, numBases = 5, type = "full")

When adding transcription direction information;
G <- readMPFile(inputFile, numBases = 5, trDir = TRUE)

}

