% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readFile.R
\name{readMutFile}
\alias{readMutFile}
\title{Read the raw mutation data.}
\usage{
readMutFile(infile, numBases = 3, trDir = FALSE, type = "independent")
}
\arguments{
\item{infile}{the path for the input file for the mutation data.}

\item{numBases}{the number of upstream and downstream flanking bases
(including the mutated base) to take into account.

VariantAnnotation::VRanges}
}
\description{
Read the raw mutation data.
}

