% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateTheta_NormalizedC}
\alias{updateTheta_NormalizedC}
\title{Update the auxiliary parameters theta and normalize them so that the summation of each group sums to 1 (E-step)}
\usage{
updateTheta_NormalizedC(vF, vQ, fdim, signatureNum, sampleNum, patternNum,
  isBackground, vF0)
}
\arguments{
\item{vF}{F (converted to a vector)}

\item{vQ}{Q (converted to a vector)}

\item{fdim}{a vector specifying the number of possible values for each mutation signature}

\item{signatureNum}{the number of mutation signatures}

\item{sampleNum}{the number of cancer genomes}

\item{patternNum}{the number of possible combinations of all the mutation features}

\item{isBackground}{the logical value showing whether a background mutaiton features is included or not}

\item{vF0}{a background mutaiton features}
}
\description{
Update the auxiliary parameters theta and normalize them so that the summation of each group sums to 1 (E-step)
}

