% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualizeSignature.R
\name{visPMS_full3}
\alias{visPMS_full3}
\title{visualize probabisitic mutaiton signature for the full model
with one 5' and 3' bases}
\usage{
visPMS_full3(v1 = rep(1/96, 96))
}
\arguments{
\item{v1}{a vector for mutation signature}
}
\description{
Generate visualization of mutation signatures for the model with
  substitution patterns and one 5' and 3' bases represented by the full
  representation.
}

