% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualizeSignature.R
\name{visPMS_ind5}
\alias{visPMS_ind5}
\title{visualize probabisitic mutaiton signature for the independent model
with two 5' and 3' bases}
\usage{
visPMS_ind5(vF = matrix(0, 5, 6))
}
\arguments{
\item{vF}{a matrix for mutation signature}
}
\description{
Generate visualization of mutation signatures for the model with
  substitution patterns and two 5' and 3' bases represented by the indepenent
  representation.
}

