% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{visPMS_ind5_dir}
\alias{visPMS_ind5_dir}
\title{visualize probabisitic mutaiton signature for the independent model
with two 5' and 3' bases and transcription direction}
\usage{
visPMS_ind5_dir(vF = matrix(0, 6, 6))
}
\arguments{
\item{vF}{a matrix for mutation signature}
}
\description{
Generate visualization of mutation signatures for the model with
  substitution patterns, two 5' and 3' bases and transcription directions represented by the indepenent
  representation.
}

